+++ src/components/CarouselCard.js
@@ -3,11 +3,11 @@ import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import { Link as RouterLink} from 'react-router-dom';
 import CardActionArea from '@material-ui/core/CardActionArea';
-//import CardActions from '@material-ui/core/CardActions';
+import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
-//import { Route } from "react-router-dom";
+import { Route } from "react-router-dom";
 
 const useStyles = makeStyles({
   root: {
@@ -20,9 +20,9 @@ const useStyles = makeStyles({
   },
 });
 
-export default function CarouselCard(props/*, match*/) {
+export default function CarouselCard(props, match) {
   const classes = useStyles();
-  const {title, description, img, id/*, pageData*/} = props;
+  const {title, description, img, id, pageData} = props;
 
     /* Create an array of `<li>` items for each product */
   // const linkList = productsData.map((product) => {
@@ -36,7 +36,7 @@ export default function CarouselCard(props/*, match*/) {
   return (
     <div>
       <Card className={classes.root}>
-        <CardActionArea component={RouterLink} to={`/${id}`}>
+        <CardActionArea component={RouterLink} to={`/actvity/${id}`}>
           <CardMedia
             className={classes.media}
             image={img}
